+++ Authoring/NavigationPortal.cs
@@ -36,7 +36,7 @@ namespace CustomNavigation.Authoring
 
         private void OnDrawGizmos()
         {
-            if (!NavigationHighlightSettings.Enabled)
+            if (!NavigationHighlightSettings.SourcesEnabled)
             {
                 return;
             }
@@ -46,7 +46,7 @@ namespace CustomNavigation.Authoring
 
         private void OnDrawGizmosSelected()
         {
-            if (NavigationHighlightSettings.Enabled)
+            if (NavigationHighlightSettings.SourcesEnabled)
             {
                 return;
             }