+++ Authoring/NavigationTestPoint.cs
@@ -35,7 +35,7 @@ namespace CustomNavigation.Authoring
 
         private void OnDrawGizmos()
         {
-            if (!NavigationHighlightSettings.Enabled)
+            if (!NavigationHighlightSettings.SourcesEnabled)
             {
                 return;
             }
@@ -45,7 +45,7 @@ namespace CustomNavigation.Authoring
 
         private void OnDrawGizmosSelected()
         {
-            if (NavigationHighlightSettings.Enabled)
+            if (NavigationHighlightSettings.SourcesEnabled)
             {
                 return;
             }